+++ src/App.js
@@ -17,7 +17,7 @@ import NotFoundPage from "./components/PageNotFound/PageNotFound";
 import GoogleTranslate from "./components/GoogleTranslate";
 import CryptoConverter from "./pages/CryptoConverter";
 import BitcoinRainbowChart from "./pages/BitCoinRainBow/BitcoinRainbowChart";
-import EthereumRainbowChart from "./pages/EthereumRainbowChart";
+import EthereumRainbowChart from "./pages/EthereumRainbow/EthereumRainbowChart";
 import Signup from "./pages/Signup";
 import Login from "./pages/Login";
 import { I18nextProvider } from "react-i18next";
